requirements.txt
python-telegram-bot==13.0
pyrogram[fast]==0.16.0
praw
pytz
peewee
toml
requests
Pillow
imgurpython
youtube-dl
apscheduler
